[fix] new handleScope after master down

+++ framework/core/src/python/kungfu/yijinjing/utils.py
@@ -10,6 +10,8 @@ def prune_layout_files(base_dir, layout, mode):
     for file in filter(os.path.isfile, glob.glob(search_path)):
         if "renderer-app" in file and layout == "log":
             continue
+        if "cli-" in file and layout == "log":
+            continue
         try:
             os.remove(file)
         except:
